Prevent bugs in config files without final newline
By ensuring a newline before the injected snippet.

I got reports outside of github from zsh users having issues with g
and the reason was that the snippet was appended to the last line of
their config file because it lacked a final newline; probably after
they manually updated such file.