Add support for listing and installing unstable versions
This also adds a big performance boost when loading remote versions.
This is because we switched the versions source from the golang.org
download page to the web UI of the git project hosted at
googlesource.com

https://go.googlesource.com/go/+refs   25 KB   59 ms
https://golang.org/dl/                341 KB  371 ms

This was not the actual reason for the switch, but the fact that the
golang.org download page doesn't list any previuos unstable versions.