Add handling of misconfigured $GOPATH/bin
`g` now throws when `$GOPATH/bin` is not listed in the `$PATH`, since
it is crutial for many commands that depend on the go binaries being
called with `command` and `g` itself for self-upgrade.

Also handle installation failing when trying to link the go binaries
if the `$GOPATH/bin/` dir doesn't exists, which is usually handled by
g-install, but nontheless found missing while doing some testing
without g-install.