Improve self-upgrade script
Before this change, often the g script would be overriten by g-install
and the shell would keep reading the file at the point it left the
buffer at. In some shells this might not happen since the whole file
would be buffered, but I've seen this happening on them all.

So now it is creating a copy of itself on a temp dir and using exec to
change the excution process, passing over the same commands.