Improve previous installation detection on g-install
I wrongly assumed in the past that `command <x>` would only lookup for
binaries in the $PATH, but it seems that in bash and zsh it actually
takes into account aliases and those take precedence.

To not handle the complexity of listing all possible resolutions for
the `g` command on diff "POSIX" shells (whatever is linked as /bin/sh)
we will assume that if `g` was installed and being managed by
`g-install`, the binary is always going to live at $GOPATH/bin/g.

This should simplify many things. If the user didn't use `g-install`,
we should and can not upgrade it for them.