Fix g not activating selected version when system one is available
activate_version would check for the output of `go version` to be
different from the selected one before proceeding. This would be
a nice little performance boost in for such usecase, but relying
on such check makes it prone to errors during g-install process.

This was discovered while running tests in newer versions of
Github Actions on which it had the latest version of go
pre-installed.